START84_1381_Rating
Take two initializer a = 1 and b = 0. then take two counter ans1 = 0 and ans2 = 0. If a == '1' increase ans1 by 1 similarly if b == '1' increase ans2 by 1. For each time upgrade a and b using X-or operation, let String[i] is not equal to 'a', Then 'a' is equal to 1. vice versa.

Finally print the max value of ans1 and ans2.