START121_1306_Rating
Try think simply, If n is equal to 1 , then it is sorted.
if n is equal to 2 , then it's easy to handle.
If n is equal to 3 , then check if is it sorted or not. if not sorted then swap first and last element and check again.

If n == 4: then, you can place any element in any index. suppose, we have ,----------
1....5....3....4 then we can make it,------------ 
1....4....3....5 ------------------------------------------ 
3....4....1....5 ------------------------------------------- 
3....5....1....4 ------------------------------------------- 
4....5....1....3 -------------------------------------------
1....5....4....3 -------------------------------------------
1....3....4....5 which is sorted. So if n is greater or equal to 4 then the ans is "YES"