START47_1390_Rating
let's consider the test case:-------------------------- 3---------------------------------------------------------------
1 9 1---------------------------------------------------------- 
for this test case we can't decrease 9 to 1. because we have the ability to increase for a student. So, we can increase first 1 to 9. then the array is , {9, 9, 1} then we have to increase last 1 also , then the final array is, {9, 9, 9} and the answer is 2. (not 1)