START79_1504_Rating
Let's consider the example below, 2-------------------------------------------------------------- 
2 3----------------------------------------------------------- 
4 3----------------------------------------------------------- 
In the first test case, the difference between 2 and 3 is 1, the minimum value is 2 that is even. so the winner is "CHEFINA". ---------------- In the next test case, difference between the piles is 1 and the minimum value is 3 that is odd. so the winner is "CHEF".----------------------

Main approch :
If the difference between the piles is 1 or 0 then check the parity and make sure the winner. otherwise , whose piles is more, he is winner.