START22_1393_Rating
let's consider the example below, 
3
32
7 
36

Now for the first test case the largest integer sequence is,
 A = [ 16, 8, 4, 2, 1] and the another is , 
S = [ 16, 24, 28, 30, 31]. the sequence A is made of integers which is the division of previous element of the sequence by 2. So ans is 5.

In the second test case, 7 is an odd number so we can't divide it by 2 that's why the ans is 0.

In the last test case, the largest integer sequence is, 
A = [ 18, 9] and the another sequence is , 
S = [ 18, 27]. the last element of A is 9 which is odd. So we can't get more elements and the expected ans is 2.