START122_825_Rating
First of all, you have to count the point (run + (wicket * 20)) of a single player and then you have to check, is it max value or not? if it the point is larger the previous max value, then mark the index of the player as a answer.

After testing all 22 player, print answer.