START122_1723_Rating
Let's consider the test case below, n = 43223. the binary representation is, 1010100011010111 formally we have to make it as like as,,,,,1100000000000000 because if all 1's or 0's gets together then there are no chance to find 101 subsequence. So the main approach is that , we have to find 1st 0 between two 1's, then we have to change the value as 1 and after that all are zero. finally we have calculate the difference of the numbers. for example, 

....................1100000000000000................................................ (43223)........1010100011010111...............................................
__________________________________________________
difference is,......1011100101001 (5929)..................................

So the answer is 5929.