START33_1420_Rating
To maximise the distinct integer, you have to generate first natural numbers as much as possible . This is our main approach. So to do this,

First of all we have to sum all the values of provided array, then we have to find the value that is suited for the condition below, (n * (n + 1)) / 2 <= sum_of_array

here the max value of n is the answer.