START128_1289_Rating
we have to transform the give array to [0,1,2,3,4..........n-1] So, to do this we have to sort the given array in ascending order and calculate the difference of each element. For example let's consider the array below, given array: [8,3,0,0,6]

sorted Array:------------0 , 0 , 3 , 6 , 8 ----------------------------------- expected Array:--------0 , 1 , 2 , 3 , 4 -----------------------------------

differences: -------------0 , 1 , 1 , 3 , 4 -----------------------------------

The total differences is 0+1+1+3+4 = 9. So the ans is 9.

NOTE: Don't forget to take long long int or int64_t