START91_1486_Rating
Try think simply. WE know,

A number of integer makes odd sum, iff total number of odd elements is odd. for example, a partition [1, 2, 7, 5] there are three odd numbers 1, 7, 5 and 3 is an odd number. So the sum of the partition will be odd.

For this problem, if we ensure that each partition will contain odd number of elements then the ans is "YES" else ans is "NO".

NOTE: don't forget to use long long int or int64_t