START59_1488_Rating
The main approach is that if the maximum value of Array C exist at the first position or size of C == 1 Then the ans is -1. otherwise split the Array C into two part following the process,

 Let C = [3, 1, 2, 7, 4, 6, 5].

 take first element that is 3 and check the others that is larger then 

we find 7 is larger. So we will split the Array into, .............. 
A = [3,1, 2] and B = [7,4,6,5];

 print size of(A) then Array A and also print size of(B) then Array B;