START90_1488_Rating
To make a string S anti-palindrome we need at most 2 operations. So the ans can be 0, 1, or 2. The conditions below,

ANS = 2 ::-- Let S = "aaaaa", it contains only one character (a) and the size of S(5) is odd. Now if we change one character it will "aabaa" that is palindrome. So need one more change, it can be "aabca". Now it is anti-palindrome and the ans is 2.

ANS = 0 ::-- Let S = "aaabbb" it contains If a string contains three distinct character 'a' & 'b' and total odd character frequency is 2. We can't rearrange it as a palindromic string. So we can say if there exist at least 2 odd number of character frequency in a string then the ans is 0.

ANS = 1 ::-- If the above two conditions aren't match then the ans is 1.