START129_1062_Rating
There are only 1 & 2 in the array so all time the product of the elements will be even or 1. Now let's see bellow , if we have some numbers of 2's then the products are,

possible x^8 ------------Total 2's ---------------------product --------- 1^8--------------------------- 0 -----------------------------------1-------------
 2^8 --------------------------- 8 ---------------------------------256 ---------- 3^8 -- NO such array possible(product odd) ----- 6561 --------- 
4^8 ------------ (2^8 X 2^8) (8+8= 16) ------------------65536 -------- 
5^8 -- NO such array possible(product odd) ----390625 ------- 
6^8 -- Not possible(2^8 X 3^8) array needs 3 --1689616------- 
8^8 ----------(2^8 X 2^8 X 2^8) (8+8+8=24) ------16777216-------

So we can see that if total number of 2's of the array is divisible by 8 then we can convert it's product as X^8.